+++ components/QuickLogModal.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const QuickLogModal = () => null;
+
+export default QuickLogModal;
